requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
requests==2.31.0
pynacl==1.5.0
azure-identity==1.15.0
azure-keyvault-keys==4.8.0
pytest==7.4.3
